+++ packages/components/CHANGELOG.md
@@ -14,6 +14,17 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0), 
 
 ## [Unreleased]
 
+### Added
+
+- `Styles`: theme for `tailwindcss` including utilities available under
+  `import { … } from '@porsche-design-system/components-{js|angular|react|vue}/tailwindcss';`
+  ([#3849](https://github.com/porsche-design-system/porsche-design-system/pull/3849))
+
+### Changed
+
+- Partials: `getInitialStyles` removed default styles for `h1`, `h2`, `h3`, `h4`, `h5`, `h6`, `p`, `b` and `strong`
+  ([#3849](https://github.com/porsche-design-system/porsche-design-system/pull/3849))
+
 #### Fixed
 
 - `Multi Select`: enable full a11y compliance
@@ -28,8 +39,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0), 
 - `Canvas`: stealing focus from `Text Field Wrapper`
   ([#3843](https://github.com/porsche-design-system/porsche-design-system/pull/3843))
 
-
-### [3.28.0-rc.1] - 2025-04-14
+## [3.28.0-rc.1] - 2025-04-14
 
 ### Fixed